Task7 ajax

+++ app/controllers/answers_controller.rb
@@ -2,26 +2,25 @@ class AnswersController < ApplicationController
   before_action :authenticate_user!
 
   expose :answer
-  expose :question , -> { Question.find(params[:question_id]) }
+  expose :question, -> { Question.find(params[:question_id]) }
 
   def create
-    @exposed_answer = question.answers.new(answer_params)
-    answer.author = current_user
+    @exposed_answer = question.answers.create(answer_params.merge(author_id: current_user.id))
+  end
 
-    if answer.save
-      redirect_to question, notice: 'Your answer successfully created.'
-    else
-      render 'questions/show'
-    end
+  def update
+    answer.update(answer_params) if current_user.author?(answer)
+    @exposed_question = answer.question
   end
 
   def destroy
-    if current_user.author?(answer)
-      answer.destroy
-      redirect_to answer.question, notice: 'Your answer has been deleted.'
-    else
-      redirect_to answer.question, notice: 'You are not the author of the answer.'
-    end
+    answer.destroy if current_user.author?(answer)
+  end
+
+  def mark_best
+    answer = Answer.find(params[:id])
+    answer.mark_best if current_user.author?(answer.question)
+    @exposed_question = answer.question
   end
 
   private
